:recycle: refactor: 패키지 네이밍 수정
