+++ semo-batch/src/main/java/sandbox/semo/batch/repository/mapper/SessionDataRowMapper.java
@@ -5,8 +5,8 @@ import java.sql.SQLException;
 import java.time.LocalDateTime;
 import lombok.RequiredArgsConstructor;
 import org.springframework.jdbc.core.RowMapper;
-import sandbox.semo.domain.collection.entity.SessionData;
-import sandbox.semo.domain.collection.entity.SessionDataId;
+import sandbox.semo.domain.monitoring.entity.SessionData;
+import sandbox.semo.domain.monitoring.entity.SessionDataId;
 import sandbox.semo.domain.device.entity.Device;
 
 @RequiredArgsConstructor
